refactor: remove programatic way of intializing; add onClose handler
