+++ stories/pages/CaseStudies.stories.tsx
@@ -3,8 +3,8 @@ import { storiesOf } from '@storybook/react';
 import { NextRouter } from 'next/router';
 import { RouterContext } from 'next/dist/shared/lib/router-context';
 
-import CaseStudiesPage from '../../pages/case-studies';
-import CaseStudyDetailPage from '../../pages/case-studies/[caseStudyId]';
+import CaseStudiesPage from '../../pages/case-studies/index.page-wip';
+import CaseStudyDetailPage from '../../pages/case-studies/[caseStudyId].page-wip';
 
 const mockRouter = {
   pathname: '/case-studies/[caseStudyId]',
